refactor: move helpers into helpers file
